SplitViewController & request lists
